organize user controller and auth controller
