+++ controllers/v1/auth_controller/check_credential.go
@@ -1,4 +1,4 @@
-package v1
+package auth_controller
 
 import (
 	"flop/config/database"
