+++ controllers/v1/auth_controller/user_sign_up.go
@@ -1,4 +1,4 @@
-package v1
+package auth_controller
 
 import (
 	"flop/helper/api_response"
