+++ controllers/v1/user_controller/update_email.go
@@ -1,4 +1,4 @@
-package v1
+package user_controller
 
 import (
 	"flop/helper/api_response"
